Задание 5. Модуляция
